update to admin [deploy:Production]

+++ includes/admin.php
@@ -70,8 +70,8 @@ END;
                             <li class="btn waves-effect waves-light color-secondary--background">
                                 <a href="/admin/process/{$CSRFtoken}/unblock/{$user['id']}/{$class}/">Unblock Password</a>
                             </li>
-                            <li class="btn waves-effect waves-light color-secondary--background">
-                                <a href="/admin/process/{$CSRFtoken}/delete/{$user['id']}/{$class}/">Delete User</a>
+                            <li class="btn waves-effect waves-light color-secondary--background disabled">
+                                <a href="/admin/process/{$CSRFtoken}/delete/{$user['id']}/{$class}/" onclick="return confirm('Weet je het zeker?')">Delete User</a>
                             </li>
                         </ul>
                     </div>
